Split dapi pricing mt generation script into two steps
